+++ default.nix
@@ -37,14 +37,14 @@ let
 in
 rec {
 
-  blockfrost-backend =
+ blockfrost-backend-ryo =
     with pkgs.lib;
     let
       src = cleanSource ./.;
       project = pkgs.callPackage ./yarn-project.nix { nodejs = pkgs.nodejs-16_x; } { inherit src; };
     in
     project.overrideAttrs (oldAttrs: rec {
-      name = "blockfrost-backend";
+      name = "blockfrost-backend-ryo";
       version = packageJSON.version;
 
       buildInputs = [
@@ -63,8 +63,8 @@ rec {
         ${nodePackages.pm2}/bin/pm2 delete all; \
            ${nodePackages.pm2}/bin/pm2 start \
            $out/libexec/source/dist/server.js \
-           --interpreter=${pkgs.nodejs-16_x}/bin/node --node-args="--max-http-header-size=32768" \
-           --max-memory-restart 1500M \
+           --interpreter=${pkgs.nodejs-16_x}/bin/node --node-args="\''${BLOCKFROST_NODE_ARGS:-"--max-http-header-size=32768"}" \
+           --max-memory-restart \''${BLOCKFROST_MAX_MEMORY_RESTART:-"1500M"} \
            -i max --time --no-daemon
         EOF
         chmod +x $out/bin/${name}
@@ -90,7 +90,7 @@ rec {
     sha256 = "1sa64g9w2dcw890d51c5xdqnav29dh7fzzvyhhwwigq7j5vinx3r";
   };
 
-  blockfrost-backend-test-mainnet = makeTest rec {
+ blockfrost-backend-ryo-test-mainnet = makeTest rec {
 
     name = "blockfrost-backend-test-mainnet";
 
@@ -99,12 +99,12 @@ rec {
       # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
       virtualisation.memorySize = 4096;
       # Backend service
-      systemd.services.blockfrost-backend-mainnet = {
+      systemd.services.blockfrost-backend-ryo = {
         wantedBy = [ "multi-user.target" ];
-        script = "${blockfrost-backend}/bin/blockfrost-backend";
+        script = "${blockfrost-backend-ryo}/bin/blockfrost-backend-ryo";
         environment = {
           # Use config file from repository
-          NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend}/libexec/source/config/mainnet.ts";
+          NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-ryo}/libexec/source/config/mainnet.ts";
           /*
             # Use this if you want to override config/default.ts
             NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-test-config}";
@@ -116,15 +116,15 @@ rec {
 
     testScript = ''
       start_all()
-      machine.wait_for_unit("blockfrost-backend-mainnet.service")
+      machine.wait_for_unit("blockfrost-backend-ryo.service")
       machine.wait_for_open_port(3000)
       machine.succeed(
-          "${pkgs.yarn}/bin/yarn set version berry && cd ${blockfrost-backend}/libexec/source && ${pkgs.yarn}/bin/yarn test-integration:mainnet"
+          "${pkgs.yarn}/bin/yarn set version berry && cd ${blockfrost-backend-ryo}/libexec/source && ${pkgs.yarn}/bin/yarn test-integration:mainnet"
       )
     '';
   };
 
-  blockfrost-backend-test-preview = makeTest rec {
+  blockfrost-backend-ryo-test-preview = makeTest rec {
 
     name = "blockfrost-backend-test-preview";
 
@@ -133,12 +133,12 @@ rec {
       # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
       virtualisation.memorySize = 4096;
       # Backend service
-      systemd.services.blockfrost-backend-preview = {
+      systemd.services.blockfrost-backend-ryo = {
         wantedBy = [ "multi-user.target" ];
-        script = "${blockfrost-backend}/bin/blockfrost-backend";
+        script = "${blockfrost-backend-ryo}/bin/blockfrost-backend-ryo";
         environment = {
           # Use config file from repository
-          NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend}/libexec/source/config/preview.ts";
+          NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-ryo}/libexec/source/config/preview.ts";
           /*
             # Use this if you want to override config/default.ts
             NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-test-config}";
@@ -150,10 +150,42 @@ rec {
 
     testScript = ''
       start_all()
-      machine.wait_for_unit("blockfrost-backend-preview.service")
+      machine.wait_for_unit("blockfrost-backend-ryo.service")
       machine.wait_for_open_port(3000)
       machine.succeed(
-          "${pkgs.yarn}/bin/yarn set version berry && cd ${blockfrost-backend}/libexec/source && ${pkgs.yarn}/bin/yarn test-integration:preview"
+          "${pkgs.yarn}/bin/yarn set version berry && cd ${blockfrost-backend-ryo}/libexec/source && ${pkgs.yarn}/bin/yarn test-integration:preview"
+      )
+    '';
+  };
+
+  blockfrost-backend-ryo-test-preprod = makeTest rec {
+
+    machine = {
+      # We have to increase memsize, otherwise we will get error:
+      # "Kernel panic - not syncing: Out of memory: compulsory panic_on_oom"
+      virtualisation.memorySize = 4096;
+      # Backend service
+      systemd.services.blockfrost-backend-ryo = {
+        wantedBy = [ "multi-user.target" ];
+        script = "${blockfrost-backend-ryo}/bin/blockfrost-backend-ryo";
+        environment = {
+          # Use config file from repository
+          NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-ryo}/libexec/source/config/preprod.ts";
+          /*
+            # Use this if you want to override config/default.ts
+            NODE_CONFIG_RUNTIME_JSON = "${blockfrost-backend-test-config}";
+          */
+        };
+      };
+
+    };
+
+    testScript = ''
+      start_all()
+      machine.wait_for_unit("blockfrost-backend-ryo.service")
+      machine.wait_for_open_port(3000)
+      machine.succeed(
+          "${pkgs.yarn}/bin/yarn set version berry && cd ${blockfrost-backend-ryo}/libexec/source && ${pkgs.yarn}/bin/yarn test-integration:preprod"
       )
     '';
   };