+++ test/integration/tests/preview/routes.ts
@@ -1,0 +1,7 @@
+import { describe } from 'vitest';
+import { previewFixtures } from '@blockfrost/blockfrost-tests';
+import { generateTestSuite } from '../../utils';
+
+describe('Integration Tests - preview', () => {
+  generateTestSuite(previewFixtures);
+});